backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic-settings==2.3.4
slowapi==0.1.9
sqlalchemy==2.0.31
alembic==1.13.2
crewai==0.55.0
langchain-openai==0.1.17
langchain-google-genai==1.0.8
langchain-community==0.2.10
chromadb==0.5.4
redis==5.0.7
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
networkx==3.3
scikit-learn==1.5.1
numpy==1.26.4
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
